made component
maar werkt niet meer
