+++ src/components/ThemeProvider.tsx
@@ -1,0 +1,59 @@
+import React, { useState, useEffect, createContext, useContext } from "react";
+
+const ThemeContext = createContext({
+    theme: 'light',
+    toggleTheme: () => {},
+});
+
+
+export const useTheme = () => useContext(ThemeContext);
+
+const ThemeProvider = ({ children }) => {
+    const [theme, setTheme] = useState("light");
+
+    useEffect(() => {
+        const savedTheme = localStorage.getItem("theme") || "light";
+        setTheme(savedTheme);
+        applyTheme(savedTheme);
+    }, []);
+
+    const toggleTheme = () => {
+        const newTheme = theme === "light" ? "dark" : "light";
+        setTheme(newTheme);
+        localStorage.setItem("theme", newTheme);
+        applyTheme(newTheme);
+    };
+
+const applyTheme = (theme) => {
+    const root = document.documentElement;
+
+    if (theme === 'light') {
+        root.style.setProperty('--accent', 'var(--light-accent)');
+        root.style.setProperty('--accent-2', 'var(--light-accent-2)');
+        root.style.setProperty('--accent-3', 'var(--light-accent-3)');
+        root.style.setProperty('--background', 'var(--light-background)');
+        root.style.setProperty('--background-2', 'var(--light-background-2)');
+        root.style.setProperty('--background-3', 'var(--light-background-3)');
+        root.style.setProperty('--text', 'var(--light-text)');
+        root.style.setProperty('--text-hover', 'var(--light-text-hover)');
+    }
+    else if (theme === 'dark') {
+        root.style.setProperty('--accent', 'var(--dark-accent)');
+        root.style.setProperty('--accent-2', 'var(--dark-accent-2)');
+        root.style.setProperty('--accent-3', 'var(--dark-accent-3)');
+        root.style.setProperty('--background', 'var(--dark-background)');
+        root.style.setProperty('--background-2', 'var(--dark-background-2)');
+        root.style.setProperty('--background-3', 'var(--dark-background-3)');
+        root.style.setProperty('--text', 'var(--dark-text)');
+        root.style.setProperty('--text-hover', 'var(--dark-text-hover)');
+    }
+};
+
+    return (
+        <ThemeContext.Provider value={{ theme, toggleTheme }}>
+            {children}
+        </ThemeContext.Provider>
+    );
+}
+
+export default ThemeProvider;
